+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/core/model/internal/CheckoutConfiguration.kt
@@ -2,11 +2,11 @@ package com.mercadopago.sdk.android.checkout.core.model.internal
 
 import android.os.Parcelable
 import com.mercadopago.sdk.android.checkout.core.model.MPCheckoutType
-import com.mercadopago.sdk.android.checkout.core.model.MPPaymentMethod
+import com.mercadopago.sdk.android.checkout.core.model.MPPaymentMethodConfig
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
 internal data class CheckoutConfiguration(
     val checkoutType: MPCheckoutType<*>,
-    val paymentMethods: List<MPPaymentMethod>,
+    val paymentMethodConfigs: List<MPPaymentMethodConfig>,
 ) : Parcelable