+++ checkout/src/test/java/com/mercadopago/sdk/android/checkout/di/DataModuleTest.kt
@@ -8,7 +8,6 @@ import com.mercadopago.sdk.android.checkout.core.model.MPOrder
 import com.mercadopago.sdk.android.checkout.core.model.MPPayer
 import com.mercadopago.sdk.android.checkout.core.model.internal.CheckoutConfiguration
 import com.mercadopago.sdk.android.checkout.data.remote.service.CardFormService
-import com.mercadopago.sdk.android.checkout.data.remote.service.OrderService
 import com.mercadopago.sdk.android.checkout.domain.model.CardFormInitializationOutput
 import com.mercadopago.sdk.android.checkout.domain.model.MPInstallmentData
 import com.mercadopago.sdk.android.checkout.domain.model.MPPaymentData
@@ -76,14 +75,13 @@ internal class DataModuleTest {
             checkoutType = MPCheckoutType.CardTransaction(
                 MPOrder(amount = BigDecimal.TEN, payer = MPPayer(email = "")),
             ),
-            paymentMethods = emptyList(),
+            paymentMethodConfigs = emptyList(),
         )
 
         val module = module {
             includes(provideDataModule())
             single { checkoutConfiguration }
             single { mockk<CardFormService>(relaxed = true) }
-            single { mockk<OrderService>(relaxed = true) }
         }
 
         val koin = koinApplication {